refactor tests to tests/rate
